updated events.yaml and EventsController.java with courtlogs get endpoint. also updated the courtlogs tests
