--- src/integrationTest/java/uk/gov/hmcts/darts/courtlogs/controller/CourtLogsControllerTest.java
@@ -1,45 +1,0 @@
-package uk.gov.hmcts.darts.courtlogs.controller;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-
-import java.time.OffsetDateTime;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@ActiveProfiles({"intTest", "h2db"})
-@AutoConfigureMockMvc
-@SuppressWarnings({"PMD.JUnitTestsShouldIncludeAssert"})
-class CourtLogsControllerTest {
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @Autowired
-    private transient MockMvc mockMvc;
-
-    @Test
-    void courtLogsGet() throws Exception {
-
-        String time
-            = OffsetDateTime.now().toString();
-
-        MockHttpServletRequestBuilder requestBuilder = get("/courtlogs")
-            .queryParam("courthouse","String")
-            .queryParam("caseNumber","String")
-            .queryParam("startDateTime",time)
-            .queryParam("endDateTime",time)
-            .contentType(MediaType.APPLICATION_JSON_VALUE);
-        mockMvc.perform(requestBuilder).andExpect(status().isNotImplemented());
-
-    }
-}
