+++ src/integrationTest/java/uk/gov/hmcts/darts/event/controller/EventsControllerCourtLogsTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
@@ -15,6 +16,7 @@ import java.net.URI;
 import java.time.OffsetDateTime;
 import java.util.Collections;
 
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -59,6 +61,22 @@ class EventsControllerCourtLogsTest extends IntegrationBase {
                                             "some-courtroom",
                                             Collections.singletonList("some-case-number"),
                                             "some-text");
+    }
+
+    @Test
+    void courtLogsGet() throws Exception {
+
+        String time
+            = OffsetDateTime.now().toString();
+
+        MockHttpServletRequestBuilder requestBuilder = get(ENDPOINT)
+            .queryParam("courthouse","String")
+            .queryParam("caseNumber","String")
+            .queryParam("startDateTime",time)
+            .queryParam("endDateTime",time)
+            .contentType(MediaType.APPLICATION_JSON_VALUE);
+        mockMvc.perform(requestBuilder).andExpect(status().isNotImplemented());
+
     }
 
 }
